"Collision" fully replaced with "CollisionManager", CharacterMotor removed, features now owned by "features.hpp"

+++ src/core/math/rng/rng.hpp
@@ -5,6 +5,6 @@
 
 namespace rng
 {
-    float getFloat(uint16_t maxValue);
+    float getFloat(float maxValue);
     uint16_t getInt(uint16_t maxValue);
 }
